Fix bugs found by gcc 4.3.2
git-svn-id: https://svn.atnf.csiro.au/difx@884 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/difx2fits/trunk/src/time2str.c
@@ -94,7 +94,7 @@ char *srvMjd2str
  */
 {
     char *poriginal = pOutStr;  /* hold output string address for return */
-    long mjd;
+    int mjd;
     double rads;
     char tempStr[40];
